scrollsafe-inference-api-cpu/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
transformers==4.48.0
huggingface-hub==0.24.5
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.9
tqdm==4.66.5
